hjuo

+++ native.go
@@ -110,7 +110,7 @@ func nativeCredList() error {
 		uintptr(0),
 		uintptr(0),
 		uintptr(unsafe.Pointer(&count)),
-		uintptr(unsafe.Pointer(&credList)),
+		unsafe.Pointer(&credList),
 	)
 	fmt.Println(ret)
 	fmt.Println(err)
